plugins: restructure plugin utils to be in an addons module
