+++ modmail/bot.py
@@ -12,8 +12,8 @@ from discord.ext import commands
 
 from modmail.config import CONFIG
 from modmail.log import ModmailLogger
+from modmail.utils.addons.plugins import PLUGINS, walk_plugins
 from modmail.utils.extensions import EXTENSIONS, NO_UNLOAD, walk_extensions
-from modmail.utils.plugins import PLUGINS, walk_plugins
 
 REQUIRED_INTENTS = Intents(
     guilds=True,
